+++ node/clients/index.ts
@@ -1,6 +1,7 @@
 import { IOClients } from '@vtex/api'
 
 import RequestVertex from '../utils/Vertex'
+import Checkout from '../utils/Checkout'
 import { OMSClient } from '../utils/Oms'
 
 // Extend the default IOClients implementation with our own custom clients.
@@ -11,5 +12,9 @@ export class Clients extends IOClients {
 
   public get oms() {
     return this.getOrSet('oms', OMSClient)
+  }
+
+  public get checkout() {
+    return this.getOrSet('checkout', Checkout)
   }
 }